Section 25 exercises. Refactoring toward GenServer. Refactor 404 counter to use GenericServer
