+++ lib/servy/four_oh_four_counter.ex
@@ -2,71 +2,54 @@ defmodule Servy.FourOhFourCounter do
 
   @name :four_oh_four_counter
 
+  alias Servy.GenericServer
+
   # Client Interface
 
   def start do
     IO.puts "Starting FourOhFourCounter..."
-    # Ensure no process is registered before starting
-    if Process.whereis(@name) do
-      Process.unregister(@name)
-    end
-    pid = spawn(__MODULE__, :listen_loop, [%{}])
-    Process.register(pid, @name)
-    pid
-  end
-
-  def bump_count(path) do
-    send @name, {self(), :bump_count, path}
-
-    receive do
-      {:response, count} -> count
-    end
-  end
-
-  def get_count(path) do
-    send @name, {self(), :get_count, path}
-
-    receive do
-      {:response, count} -> count
-    end
-  end
-
-  def get_counts do
-    send @name, {self(), :get_counts}
-
-    receive do
-      {:response, counts} -> counts
-    end
-  end
-
-  # Server
-
-  def listen_loop(state) do
-    receive do
-      {sender, :bump_count, path} ->
-        new_state = Map.update(state, path, 1, &(&1 + 1))
-        send sender, {:response, Map.get(new_state, path)}
-        listen_loop(new_state)
-
-      {sender, :get_count, path} ->
-        count = Map.get(state, path, 0)
-        send sender, {:response, count}
-        listen_loop(state)
-
-      {sender, :get_counts} ->
-        send sender, {:response, state}
-        listen_loop(state)
-
-      unexpected ->
-        IO.puts "Unexpected message: #{inspect unexpected}"
-        listen_loop(state)
-    end
+    GenericServer.start(__MODULE__, @name, %{})
   end
 
   def stop do
-    if pid = Process.whereis(:four_oh_four_counter) do
-      Process.exit(pid, :normal)
-    end
-    :ok
+    IO.puts "Stopping FourOhFourCounter..."
+    GenericServer.stop(@name)
+  end
+
+  def bump_count(path) do
+    GenericServer.call @name, {:bump_count, path}
+  end
+
+  def get_count(path) do
+    GenericServer.call @name, {:get_count, path}
+  end
+
+  def get_counts do
+    GenericServer.call @name, :get_counts
+  end
+
+  def clear do
+    GenericServer.cast @name, :clear
+  end
+
+  # Server Callbacks
+
+  def handle_call({:bump_count, path}, state) do
+    new_state = Map.update(state, path, 1, &(&1 + 1))
+    {Map.get(new_state, path), new_state}
+  end
+
+  def handle_call({:get_count, path}, state) do
+    count = Map.get(state, path, 0)
+    {count, state}
+  end
+
+  def handle_call(:get_counts, state) do
+    {state, state}
+  end
+
+  def handle_cast(:clear, _state) do
+    IO.puts "Clearing counts..."
+    %{}
   end
 end
